pyproject.toml
[tool.poetry]
name = "binance-trading-bot"
version = "0.1.0"
description = ""
authors = ["idabblewith <109533655+idabblewith@users.noreply.github.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
requests = "^2.32.3"
python-dotenv = "^1.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
